refactor: extract pure utils to js/utils.js — single source of truth
* refactor: extract pure utils to js/utils.js — single source of truth

- Move escHtml, fmtCount, matchesTier, fixImgPaths, filterProblems,
  getPaginationSlice, TIER_NAMES, TIER_COLORS into js/utils.js as ESM exports
- js/main.js imports from ./utils.js; applyFilter/render use the shared fns
- index.html: <script type="module"> so the import statement resolves
- src/utils.js: replaced with `export * from '../js/utils.js'` — tests
  now import the same code the app runs, eliminating drift risk


* fix: use end from getPaginationSlice instead of recalculating

---------
