+++ index.html
@@ -161,7 +161,7 @@
             </div>
         </div>
 
-        <script src="js/main.js"></script>
+        <script type="module" src="js/main.js"></script>
 
         <footer class="site-footer">
             본 사이트는 비영리 보존 목적으로 운영됩니다.<br />
